+++ src/styles/Header.css
@@ -1,0 +1,102 @@
+header {
+  display: grid;
+  grid-template-columns: 70% 30%;
+  position: fixed;
+  width: 100%;
+  padding: 0.5rem;
+  z-index: 1000;
+  gap: 2rem;
+}
+
+.header-article {
+  grid-column: 1 / 2;
+  width: 100%;
+  height: 4rem;
+  color: #fff;
+  background: #86082e;
+  border-radius: 8px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 1rem;
+}
+
+header article ul {
+  display: flex;
+  gap: 1rem;
+}
+
+header article ul li {
+  font-size: 1.4rem;
+  font-weight: 700;
+  padding: 0.5rem 1rem;
+  border-radius: 8px;
+}
+
+.input-busqueda {
+  display: flex;
+  align-items: center;
+  background: #86082eb2;
+  border-radius: 8px;
+  width: 90%;
+  padding: 5px 10px;
+  height: 4rem;
+  transition: all 0.3s ease-in-out;
+}
+
+.search-input {
+  flex: 1;
+  border: none;
+  outline: none;
+  background: transparent;
+  color: white;
+  font-size: 1rem;
+  padding: 5px;
+  width: 100%;
+}
+
+.search-icon {
+  cursor: pointer;
+  transition: transform 0.2s ease-in-out;
+}
+
+.search-icon:hover {
+  transform: scale(1.2);
+}
+
+@media screen and (max-width: 1024px) {
+  header {
+    grid-template-columns: 1fr;
+    padding: 0.3rem;
+  }
+
+  .header-article {
+    flex-direction: column;
+    height: auto;
+    padding: 0.5rem;
+  }
+
+  header article ul {
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+
+  .input-busqueda {
+    margin-top: 0.5rem;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .header-article {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  header article ul li {
+    display: none;
+  }
+
+  .input-busqueda input {
+    width: 150px;
+  }
+}
